Refactor overlay to class file
