--- gc_bot.log
@@ -1,1 +1,0 @@
-2025-01-13 23:15:09,451 EXITING PROGRAM
